+++ test/unit/fuzz-tools.test.js
@@ -258,6 +258,7 @@ import * as inflamV267 from '../../lib/inflam-v267.js';
 import * as inflamV268 from '../../lib/inflam-v268.js';
 import * as metabolicV269 from '../../lib/metabolic-v269.js';
 import * as adiposityV270 from '../../lib/adiposity-v270.js';
+import * as lipidsV271 from '../../lib/lipids-v271.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -460,6 +461,7 @@ const MODULES = {
   'inflam-v268.js': inflamV268,
   'metabolic-v269.js': metabolicV269,
   'adiposity-v270.js': adiposityV270,
+  'lipids-v271.js': lipidsV271,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];